+++ front_py/articles/Row_articles.py
@@ -14,7 +14,9 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-
+from core import elements as e
+from rows import Section_row
+from Article import Article
 class Row_articles(Section_row):
     """
     Creates a lot of articles. Each article in one single row
