+++ front_py/containers/Block_container.py
@@ -14,6 +14,8 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from core import elements as e
+
 
 class Block_container(e._Block_element):
     """
