+++ front_py/containers/Section_container.py
@@ -14,3 +14,12 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from core import elements as e
+
+class Section_container(e._Section_element):
+    """
+    The same as block but section, only semantic difference
+    """
+    def __init__(self):
+        super().__init__()
+        self._add_class("container")
