+++ front_py/core/back_end/functions.py
@@ -14,13 +14,10 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-from __custom_import__ import prepare_import
-prepare_import()
-from Core_http_process import Core_http_process
 from exceptions import *
 from constants import *
 from http.server import HTTPServer
-
+from back_end import Core_http_process
 def make_name(address):
     """
     Make unique name for function from request.path
