+++ front_py/core/front_end/Page.py
@@ -1,8 +1,4 @@
 """
-This is core module of application. It contains all exceptions types, Core_meta and Core_element which is the main feature of the framework
-which makes it possible to build html tags in python objects.
-
-
 Copyright (C) 2018  Dima Lukashov github.com/DimonLuk
     
 This program is free software: you can redistribute it and/or modify
@@ -18,8 +14,6 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-from __custom_import__ import prepare_import
-prepare_import()
 from constants import *
 from Core_element import Core_element
