+++ front_py/core/front_end/elements.py
@@ -40,8 +40,6 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-from __custom_import__ import prepare_import
-prepare_import()               
 from Core_element import Core_element
