+++ front_py/defaults/Image.py
@@ -14,6 +14,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from core import elements as e
 
 class Image(e._Link_element):
     """
