+++ front_py/defaults/Paragraph.py
@@ -14,6 +14,8 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from core import elements as e
+
 class Paragraph(e._Paragraph_element):
     """
     Simple paragraph
