+++ front_py/defaults/__init__.py
@@ -14,3 +14,9 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+
+from Header import Header,Header_text
+from Image import Image
+from Link import Link
+from Paragraph import Paragraph
+from Text import Text
