+++ front_py/footers/Footer.py
@@ -14,6 +14,8 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from core import elements as e
+from rows import Container_row
 
 class Footer(e._Footer_element):
     """
