+++ front_py/footers/__init__.py
@@ -14,3 +14,4 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from Footer import Footer
