+++ front_py/rows/Block_row.py
@@ -14,7 +14,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-
+from core import elements as e
 class Block_row(e._Block_element):
     """
     Simple row
