+++ front_py/rows/Container_row.py
@@ -14,7 +14,8 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-
+from containers import Block_container
+from Block_row import Block_row
 class Container_row(Block_container):
     """
     Creates row inside a Block_container
